Rolling back changes to not use enum and move all Operators into one unified file.
